Move user router to package api/user
